feat(mypage): 마이페이지 인터뷰 응답/스타일 구조 개선
