+++ src/main/java/com/allinone/DevView/mypage/dto/MypageResponseDto.java
@@ -1,5 +1,6 @@
 package com.allinone.DevView.mypage.dto;
 
+import com.allinone.DevView.interview.dto.response.InterviewResultResponse;
 import com.allinone.DevView.user.entity.User;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -19,7 +20,7 @@ public class MypageResponseDto {
     private final String job;
     private final String careerLevel;
     private String profileImageUrl;
-    private final String selfIntroduction; //자기소개
+    private final String selfIntroduction;
 
     // 📌 계정 정보
     private final Long memberId;
@@ -31,14 +32,14 @@ public class MypageResponseDto {
     private final String grade;
 
     // 📋 상세 리스트
-    private final List<InterviewDto> interviews;
+    private final List<InterviewResultResponse> interviews;
     private final List<ScrapDto> scraps;
 
     public static MypageResponseDto from(User user,
                                          int totalInterviews,
                                          int avgScore,
                                          String grade,
-                                         List<InterviewDto> interviews,
+                                         List<InterviewResultResponse> interviews,
                                          List<ScrapDto> scraps) {
 
         String job = null;
