+++ src/main/resources/static/css/mypage.css
@@ -231,36 +231,29 @@
     margin-bottom: 12px;
     position: relative;
 }
+
 .mypage-container .interview-item .row {
     display: flex;
+    align-items: center;
     justify-content: space-between;
-    align-items: flex-start;
-    gap: 12px;
+    gap: 16px;
 }
 
-/* 칩 라인: 1번째 칩(타입/직무)만 알약, 2번째 칩(날짜)은 평문 */
 .mypage-container .chips {
     display: flex;
     gap: 8px;
     flex-wrap: wrap;
 }
+
 .mypage-container .chips .chip {
-    background: transparent;
-    padding: 0;
-    border-radius: 0;
-    color: #6b7a90;
-    font-weight: 600;
-    align-content: center;
-}
-.mypage-container .chips .chip:first-child,
-.mypage-container .chip-type,
-.mypage-container .pill-dark {
     background: #20374d !important;
     color: #fff !important;
     padding: 6px 12px;
     border-radius: 999px;
     font-weight: 700;
+    font-size: 0.9rem;
 }
+
 .mypage-container .chip-date,
 .mypage-container .date-chip,
 .mypage-container .chips .date {
@@ -270,7 +263,15 @@
     color: #6b7a90 !important;
 }
 
-/* 점수/등급 영역 */
+/* 날짜(중앙) */
+.mypage-container .interview-item .left .interview-date {
+    color: #6b7280;
+    font-weight: 500;
+    min-width: 110px;
+    white-space: nowrap;
+}
+
+/* 점수/등급 영역(행 기준) */
 .mypage-container .score-wrap {
     display: flex;
     flex-direction: column;
@@ -278,6 +279,13 @@
     gap: 8px;
     margin-left: auto;
 }
+.mypage-container .interview-item .score-wrap {
+    display: flex;
+    align-items: baseline;
+    gap: 12px;
+    white-space: nowrap;
+}
+
 .mypage-container .score {
     background: transparent;
     padding: 0;
@@ -361,8 +369,8 @@
 /* ========== 상세보기 버튼: 카드 우하단 고정 (SSR + JS 공통, 선택자 보강) ========== */
 .mypage-container .interview-item,
 .mypage-container .interview-row {
-    position: relative;           /* 절대배치 기준 */
-    padding-bottom: 64px;         /* 버튼 높이+여백만큼 공간 확보(겹침 방지) */
+    position: relative;
+    padding-bottom: 64px;
 }
 /* SSR: <a class="btn small">상세 보기</a> (직접 자식) */
 .mypage-container .interview-item > a.btn.small {
@@ -421,9 +429,8 @@
     grid-template-columns: 1fr 1fr;
     gap: 16px;
 }
-/* ==============================   자기소개 카드 스타일   ============================== */
 
-/* 자기소개 카드 */
+/* ==============================   자기소개 카드 스타일   ============================== */
 .mypage-container .self-intro-card {
     background: var(--card-bg);
     border-radius: var(--radius);
@@ -431,7 +438,6 @@
     padding: 20px;
 }
 
-/* 카드 제목 (기존 account-card와 동일한 스타일) */
 .mypage-container .self-intro-card .card-title {
     display: flex;
     align-items: center;
@@ -441,14 +447,11 @@
     font-weight: 700;
     color: #0f172a;
 }
-
 .mypage-container .self-intro-card .card-title .icon {
     width: 20px;
     height: 20px;
     display: inline-block;
-    /* 다른 카드들과 동일한 기본 색상으로 변경 */
 }
-
 .mypage-container .self-intro-card .card-title h3 {
     margin: 0;
     font-size: inherit;
@@ -469,8 +472,8 @@
     margin: 0;
     color: #334155;
     font-size: 0.95rem;
-    white-space: pre-wrap; /* 줄바꿈 유지 */
-    word-break: break-word; /* 긴 단어 줄바꿈 */
+    white-space: pre-wrap;
+    word-break: break-word;
 }
 
 /* 자기소개 안내 문구 (없을 때) */
@@ -515,49 +518,23 @@
     font-weight: 500;
     transition: background-color 0.15s ease, transform 0.1s ease;
 }
-
 .mypage-container .self-intro-card .add-intro-btn:hover {
     background: #3a9bc1;
     transform: translateY(-1px);
 }
-
 .mypage-container .self-intro-card .add-intro-btn .btn-icon {
     width: 16px;
     height: 16px;
-    filter: brightness(0) invert(1); /* 흰색으로 변경 */
-}
+    filter: brightness(0) invert(1);
 
-/* 반응형: 모바일에서 자기소개 카드 조정 */
-@media (max-width: 640px) {
-    .mypage-container .self-intro-card {
-        padding: 16px;
-    }
-
-    .mypage-container .self-intro-card .self-intro-content {
-        padding: 12px;
-        font-size: 0.9rem;
-    }
-
-    .mypage-container .self-intro-card .self-intro-empty {
-        padding: 20px 12px;
-    }
-
-    .mypage-container .self-intro-card .empty-message {
-        font-size: 0.85rem;
-    }
-
-    .mypage-container .self-intro-card .add-intro-btn {
-        font-size: 0.85rem;
-        padding: 7px 14px;
-    }
-}
 /* ==============================   반응형   ============================== */
-@media (max-width: 1024px) {
-    .mypage-container .profile-layout { grid-template-columns: 1fr; }
-}
-@media (max-width: 860px) {
-    .mypage-container .scrap-section .scrap-grid { grid-template-columns: 1fr; }
-}
 @media (max-width: 640px) {
-    .mypage-container .graph-card canvas { height: 300px !important; }
+    .mypage-container .self-intro-card { padding: 16px; }
+    .mypage-container .self-intro-card .self-intro-content { padding: 12px; font-size: 0.9rem; }
+    .mypage-container .self-intro-card .self-intro-empty { padding: 20px 12px; }
+    .mypage-container .self-intro-card .empty-message { font-size: 0.85rem; }
+    .mypage-container .self-intro-card .add-intro-btn { font-size: 0.85rem; padding: 7px 14px; }
 }
+@media (max-width: 1024px) { .mypage-container .profile-layout { grid-template-columns: 1fr; } }
+@media (max-width: 860px)  { .mypage-container .scrap-section .scrap-grid { grid-template-columns: 1fr; } }
+@media (max-width: 640px)  { .mypage-container .graph-card canvas { height: 300px !important; } }
