+++ src/main/resources/templates/mypage/mypage.html
@@ -3,18 +3,17 @@
 <head>
     <meta charset="UTF-8" />
     <title>마이페이지</title>
-    <!-- CSRF 토큰 메타 태그 추가 -->
+
     <meta name="_csrf" th:content="${_csrf.token}" />
     <meta name="_csrf_header" th:content="${_csrf.headerName}" />
 
     <link rel="stylesheet" th:href="@{/css/mypage.css}" />
     <link rel="stylesheet" th:href="@{/css/header.css}" />
     <link rel="stylesheet" th:href="@{/css/password-change.css}" />
-    <!-- Chart.js (이미 포함) -->
+
     <script th:src="@{/js/lib/chart.min.js}"></script>
-    <!-- 통합 JS -->
     <script defer th:src="@{/js/mypage.js}"></script>
-    <!-- 비밀번호 변경 JS -->
+
     <script defer th:src="@{/js/password-change.js}"></script>
 </head>
 <body>
@@ -122,20 +121,22 @@
                     <img src="/img/interviewslist.svg" class="icon" alt="나의 면접 목록"/>
                     <h3>나의 면접 목록</h3>
                 </div>
+
                 <ul id="interviewList">
                     <li th:each="interview : ${user.interviews}">
                         <div class="interview-item">
                             <div class="row">
                                 <div class="chips">
-                                    <span class="chip chip-type" th:text="${interview.jobPositionDisplayName}">직무</span>
+                                    <span class="chip chip-type"  th:text="${interview.jobPositionDisplayName}">직무</span>
                                     <span class="chip chip-level" th:text="${interview.careerLevelDisplayName}">레벨</span>
-                                    <span class="chip chip-type" th:text="${interview.interviewType}">면접타입</span>
-                                    <span class="chip chip-date"
-                                          th:text="${interview.interviewDate != null ? interview.interviewDate : ''}">YYYY-MM-DD</span>
+                                    <span class="chip chip-type"  th:text="${interview.interviewType}">면접타입</span>
                                 </div>
+
+                                <span class="interview-date"
+                                      th:text="${interview.interviewDate != null ? #strings.replace(interview.interviewDate,'-','.') : ''}"> YYYY.MM.DD </span>
                                 <div class="score-wrap">
-                                    <span class="score" th:text="${interview.score + '점'}">0점</span>
-                                    <span class="grade" th:text="${interview.grade + '등급'}">-</span>
+                                    <span class="score" th:text="${interview.totalScore}">0</span>
+                                    <span class="grade" th:text="${interview.grade + ' 등급'}">-</span>
                                 </div>
                             </div>
                             <a class="btn small" th:href="${interview.detailUrl}">상세 보기</a>
@@ -154,8 +155,6 @@
                 <ul id="scrapList" class="scrap-grid">
                     <li th:each="scrap : ${user.scraps}"
                         th:with="isExternal=${#strings.startsWith(scrap.link,'http')}">
-
-                        <!-- 링크: 컨텍스트 경로 보정(@{...}) + 외부/내부에 따라 target/rel 분기 -->
                         <a th:href="@{${scrap.link}}"
                            th:attr="target=${isExternal ? '_blank' : null},
                   rel=${isExternal ? 'noopener noreferrer' : null}"
